virt-scripts: scripts to build xfstests vms
Scripts to automatically build vm's to run xfstests on, which adds in
the storage that it needs, sets up ssh keys, ip addresses, and hosts
entries.  Scripts to install kernels onto them, and a bare config to use
with all the things you need turned on to run xfstests (on btrfs
anyway).

Signed-off-by: Josef Bacik <josef@toxicpanda.com>